// src/app/interfaces/my-profile.ts
export interface MyProfile {
    nickname: string;
    birth: string;
    gender: string;
    area: string;
}

export interface FathersProfile {
    nickname: string;
    birth: string;
    email_address: string;
}


export interface MothersProfile {
    nickname: string;
    birth: string;
    email_address: string;
}